Feat: add useGame hook
